+++ overlays/haskell.nix
@@ -531,7 +531,7 @@ final: prev: {
           final.lib.fix (project':
             let project = project' // { recurseForDerivations = false; };
             in rawProject // rec {
-              hsPkgs = (final.lib.mapAttrs (n: package':
+              hsPkgs = final.lib.mapAttrs (n: package':
                 if package' == null
                   then null
                   else
@@ -554,13 +554,13 @@ final: prev: {
                         ghc = project.pkg-set.config.ghc.package;
                       });
                     }
-                ) rawProject.hsPkgs
-                // {
+                ) (builtins.removeAttrs rawProject.hsPkgs
                   # These are functions not packages
-                  inherit (rawProject.hsPkgs) shellFor makeConfigFiles ghcWithHoogle ghcWithPackages;
-              });
+                  [ "shellFor" "makeConfigFiles" "ghcWithHoogle" "ghcWithPackages" ]);
 
             projectCoverageReport = haskellLib.projectCoverageReport (map (pkg: pkg.coverageReport) (final.lib.attrValues (haskellLib.selectProjectPackages hsPkgs)));
+
+            inherit (rawProject.hsPkgs) makeConfigFiles ghcWithHoogle ghcWithPackages shellFor;
           });
 
         cabalProject =
@@ -568,12 +568,7 @@ final: prev: {
             let
               args = { caller = "hackage-package"; } // args';
               p = cabalProject' args;
-            in p.hsPkgs // {
-              inherit (p) plan-nix index-state tool tools roots projectCoverageReport;
-              # Provide `nix-shell -A shells.ghc` for users migrating from the reflex-platform.
-              # But we should encourage use of `nix-shell -A shellFor`
-              shells.ghc = p.hsPkgs.shellFor {};
-            };
+            in p.hsPkgs // p;
 
         stackProject' =
             { src, ... }@args:
@@ -601,12 +596,7 @@ final: prev: {
             in project;
 
         stackProject = args: let p = stackProject' args;
-            in p.hsPkgs // {
-              inherit (p) stack-nix tool tools roots projectCoverageReport;
-              # Provide `nix-shell -A shells.ghc` for users migrating from the reflex-platform.
-              # But we should encourage use of `nix-shell -A shellFor`
-              shells.ghc = p.hsPkgs.shellFor {};
-            };
+            in p.hsPkgs // p;
 
         # `project'` and `project` automatically select between `cabalProject`
         # and `stackProject` (when possible) by looking for `stack.yaml` or
@@ -646,12 +636,7 @@ final: prev: {
           let
             args = { caller = "project"; } // args';
             p = project' args;
-          in p.hsPkgs // {
-              # Provide `nix-shell -A shells.ghc` for users migrating from the reflex-platform.
-              # But we should encourage use of `nix-shell -A shellFor`
-              shells.ghc = p.hsPkgs.shellFor {};
-            } // final.lib.optionalAttrs (p ? stack-nix) { inherit (p) stack-nix; }
-              // final.lib.optionalAttrs (p ? plan-nix ) { inherit (p) plan-nix;  };
+          in p.hsPkgs // p;
 
         # Like `cabalProject'`, but for building the GHCJS compiler.
         # This is exposed to allow GHCJS developers to work on the GHCJS