+++ changelog.md
@@ -1,8 +1,16 @@
 This file contains a summary of changes to Haskell.nix and `nix-tools`
 that will impact users.
 
+## Nov 26, 2020
+* Renamed `otherShells` arg for `shellFor` to `inputsFrom
+
 ## Nov 25, 2020
-* Renamed `otherShells` arg for `shellFor` to `inputsFrom`.
+* The `shellFor` `makeConfigFiles` `ghcWithHoogle` and `ghcWithPackages`
+  functions have been removed from `project.hsPkgs`.  Instead access
+  them from `project` itself (e.g. change `p.hsPkgs.shellFor` to `p.shellFor`).
+* The reflex-platform like `project.shells.ghc` has been removed.
+  If needed, add something like `p // { shells.ghc = p.shellFor {} }`
+  to `shell.nix`.
 
 ## Nov 24, 2020
 * Added `${targetPrefix}cabal` wrapper script for running cross